Add powerups and enemies

+++ game/source/Components/MovementComponents.cpp
@@ -10,7 +10,7 @@ void Components::MRU::Update(float deltaTime, Transform& t)
 void Components::CyclicMRU::Update(float deltaTime, Transform& t)
 {
 	angle += speed * deltaTime;
-	angle = static_cast<int>(angle) % 360;
+	angle = static_cast<float>(static_cast<int>(angle) % 360);
 	auto factor = std::cos(angle * Geometry::DEG_TO_RAD);
 	t.position = initialPosition + distance * factor;
 }
